[fix]: move view to api
