+++ situation_prediction/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from situation_prediction.views import SituationPredictionView, SituationPredictionHistoryView
+from .api.situation_prediction import SituationPredictionView, SituationPredictionHistoryView
 
 urlpatterns = [
     path('value/', SituationPredictionView.as_view()),
